Bug 695243 - Cast the 32-bit integers to 16-bit ones explicitly in order to fix clang builds; r=benwa
This is fine since the dimensions of a plugin cannot be more than USHRT_MAX.

+++ dom/plugins/ipc/PluginInstanceChild.cpp
@@ -474,7 +474,8 @@ CAUpdate(NPP npp, uint32_t timerID) {
 void
 PluginInstanceChild::Invalidate()
 {
-    NPRect windowRect = {0, 0, mWindow.height, mWindow.width};
+    NPRect windowRect = {0, 0, uint16_t(mWindow.height),
+        uint16_t(mWindow.width)};
 
     InvalidateRect(&windowRect);
 }
